+++ R/read_skyline.R
@@ -19,6 +19,7 @@
 #' fm <- data.frame(Processing_name = c("PGE2", "PGD2"), Report = "YES",
 #'                  row.names = c("PGE2", "PGD2"))
 #' read_skyline(f, "skyline_data", fm)
+#' @family workflow steps
 #' @export
 read_skyline = function(xlsx_path, data_tab_names = "skyline_data",
                         feature_meta, signal_filter = FALSE){